+++ src/main/java/com/aisip/OnO/backend/admin/service/AdminService.java
@@ -1,6 +1,6 @@
 package com.aisip.OnO.backend.admin.service;
 
-import com.aisip.OnO.backend.fileupload.service.FileUploadService;
+import com.aisip.OnO.backend.util.fileupload.service.FileUploadService;
 import com.aisip.OnO.backend.folder.service.FolderService;
 import com.aisip.OnO.backend.problem.service.ProblemService;
 import com.aisip.OnO.backend.user.entity.User;
