--- src/main/java/com/aisip/OnO/backend/fcm/dto/FcmTokenRequestDto.java
@@ -1,6 +1,0 @@
-package com.aisip.OnO.backend.fcm.dto;
-
-public record FcmTokenRequestDto (
-        String token
-){
-}
