--- src/main/java/com/aisip/OnO/backend/fcm/dto/NotificationRequestDto.java
@@ -1,11 +1,0 @@
-package com.aisip.OnO.backend.fcm.dto;
-
-import java.util.Map;
-
-public record NotificationRequestDto (
-        String token,
-        String title,
-        String body,
-        Map<String, String> data
-){
-}
