--- src/main/java/com/aisip/OnO/backend/quartz/AutowiringSpringBeanJobFactory.java
@@ -1,24 +1,0 @@
-package com.aisip.OnO.backend.quartz;
-
-import org.quartz.spi.TriggerFiredBundle;
-import org.springframework.beans.factory.config.AutowireCapableBeanFactory;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.scheduling.quartz.SpringBeanJobFactory;
-
-public class AutowiringSpringBeanJobFactory extends SpringBeanJobFactory implements ApplicationContextAware {
-
-    private AutowireCapableBeanFactory beanFactory;
-
-    @Override
-    public void setApplicationContext(ApplicationContext context) {
-        this.beanFactory = context.getAutowireCapableBeanFactory();
-    }
-
-    @Override
-    protected Object createJobInstance(TriggerFiredBundle bundle) throws Exception {
-        Object job = super.createJobInstance(bundle);
-        beanFactory.autowireBean(job); // ★ DI 핵심
-        return job;
-    }
-}
