--- src/main/java/com/aisip/OnO/backend/quartz/QuartzConfig.java
@@ -1,25 +1,0 @@
-package com.aisip.OnO.backend.quartz;
-
-import org.quartz.spi.JobFactory;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.quartz.SchedulerFactoryBean;
-
-@Configuration
-public class QuartzConfig {
-
-    @Bean
-    public JobFactory jobFactory(ApplicationContext applicationContext) {
-        AutowiringSpringBeanJobFactory jobFactory = new AutowiringSpringBeanJobFactory();
-        jobFactory.setApplicationContext(applicationContext);
-        return jobFactory;
-    }
-
-    @Bean
-    public SchedulerFactoryBean schedulerFactoryBean(JobFactory jobFactory) {
-        SchedulerFactoryBean factory = new SchedulerFactoryBean();
-        factory.setJobFactory(jobFactory); // 여기에 설정
-        return factory;
-    }
-}
