+++ src/main/java/com/aisip/OnO/backend/util/fcm/dto/FcmTokenRequestDto.java
@@ -1,0 +1,6 @@
+package com.aisip.OnO.backend.util.fcm.dto;
+
+public record FcmTokenRequestDto (
+        String token
+){
+}
