+++ src/main/java/com/aisip/OnO/backend/util/fcm/dto/FcmTokenResponseDto.java
@@ -1,0 +1,24 @@
+package com.aisip.OnO.backend.util.fcm.dto;
+
+import com.aisip.OnO.backend.util.fcm.entity.FcmToken;
+import lombok.AccessLevel;
+import lombok.Builder;
+
+import java.time.LocalDateTime;
+
+@Builder(access = AccessLevel.PRIVATE)
+public record FcmTokenResponseDto(
+        Long id,
+        Long userId,
+        String token,
+        LocalDateTime createdAt
+) {
+    public static FcmTokenResponseDto from(FcmToken fcmToken) {
+        return FcmTokenResponseDto.builder()
+                .id(fcmToken.getId())
+                .userId(fcmToken.getUserId())
+                .token(fcmToken.getToken())
+                .createdAt(fcmToken.getCreatedAt())
+                .build();
+    }
+}
