+++ src/main/java/com/aisip/OnO/backend/util/fcm/dto/NotificationRequestDto.java
@@ -1,0 +1,11 @@
+package com.aisip.OnO.backend.util.fcm.dto;
+
+import java.util.Map;
+
+public record NotificationRequestDto (
+        String token,
+        String title,
+        String body,
+        Map<String, String> data
+){
+}
