+++ src/main/java/com/aisip/OnO/backend/util/fcm/repository/FcmTokenRepository.java
@@ -1,0 +1,16 @@
+package com.aisip.OnO.backend.util.fcm.repository;
+
+import com.aisip.OnO.backend.util.fcm.entity.FcmToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface FcmTokenRepository extends JpaRepository<FcmToken, Long> {
+
+    boolean existsByToken(String token);
+
+    Optional<FcmToken> findByToken(String token);
+
+    List<FcmToken> findAllByUserId(Long userId);
+}
