+++ src/main/java/com/aisip/OnO/backend/util/webhook/DiscordWebhookPayload.java
@@ -1,0 +1,17 @@
+package com.aisip.OnO.backend.util.webhook;
+
+import java.util.List;
+
+public record DiscordWebhookPayload(
+        String content,
+        List<Embed> embeds
+) {
+    public record Embed(
+            String title,
+            String description,
+            String timestamp,
+            List<Field> fields
+    ) {
+        public record Field(String name, String value, boolean inline) {}
+    }
+}
